Handle most types of git remote urls
The old behaviour of url normalization depended a lot on the specifics
of GitHub urls, which made it error out on other source forge urls
(particularly sourcehut in my case).

This code makes fewer assumptions about how your urls are formatted and
picks apart the relevant sections of the url and uses those for the uri
pieces. It still can't handle -every- possible corner case, but it ought
to deal with most sane urls.