[FIX] Resolve dependencies for `--path`-only projects
I have a proper `requies` specification in the `htsparse`, which lists
`hmisc` as a dependency. But `hmisc` itself is specified using `--path:`
in the directory above - as a result `nimph` failed to resolve dependencies
with `"no path to ... as hmisc"`. I'm not sure if this is a correct way to
solve the issue, but yielding all found projects seems to work.